(update) update all package and doc

+++ database/migrations/2026_03_09_102541_delete_image.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (! Schema::hasTable('projects')) {
+            return;
+        }
+
         Schema::table('projects', function (Blueprint $table) {
             $table->dropColumn('image');
         });
